Add User model

+++ controllers/admin/student.js
@@ -1,9 +1,11 @@
 const mongoose = require('mongoose');
 
+//model
 const Course = require('../../models/courses.js');
 const Admin = require('../../models/admin.js');
 const Instruction = require('../../models/instruction.js');
 const Student = require('../../models/student.js');
+const User = require('../../models/users.js');
 
 //get full list of student
 exports.getFullStudent = async (req, res, next) => {
@@ -50,6 +52,14 @@ exports.createStudent = async (req, res, next) => {
             course: newCourse
         });
 
+        //save to User model
+        const newUser = new User({
+            email: newEmail,
+            password: newPassword,
+            role: "Student"
+        })
+
+        await newUser.save();
         //save to Student Model
         await newStudent.save();
 
@@ -73,15 +83,30 @@ exports.editStudent = async (req, res, next) => {
         const newName = req.body.name;
         const newEmail = req.body.email;
         const newPassword = req.body.password;
+
         const updateStudent = await Student.findById(studentId);
         if (!updateStudent) {
             return res.status(404).json({ "message": "Not found!" });
         }
+
+        const currentEmail = updateStudent.email;
+        const currentUser = await User.findOne({ email: currentEmail.trim() });
+
+        if (!currentUser) {
+            return res.status(404).json({ "message": "User not found!" });
+        }
+
         updateStudent.name = newName;
         updateStudent.email = newEmail;
         updateStudent.password = newPassword;
+
         //save to Student Model
         await updateStudent.save();
+
+        currentUser.email = newEmail;
+        currentUser.password = newPassword;
+        await currentUser.save();
+
         res.status(200).json({ "message": "created", "statusCode": 200, "updateStudent": updateStudent });
     } catch (error) {
         console.log(error);
@@ -98,6 +123,14 @@ exports.deleteStudent = async (req, res, next) => {
         if (!currentStudent) {
             return res.status(404).json({ "message": "Not found", "statusCode": 404 })
         }
+
+        const currentEmail = currentStudent.email;
+        const currentUser = await User.findOne({ email: currentEmail });
+        if (!currentUser) {
+            return res.status(404).json({ "message": "Not found", "statusCode": 404 });
+        }
+
+        await User.deleteOne({ email: currentUser.email });
 
         await Student.findByIdAndDelete(studentId);
